+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 ## [unreleased]
 
 ### Fixed
-Taxon Pages Biological Associations returns no results
+- Taxon Pages Biological Associations returns no results
+- error on /api/v1/dwc_occurrences
 
 ## [0.62.0] - 2026-05-08